Separated classes into their own respective files
