+++ Decorator.Demo/MessageTypes/Chat.cs
@@ -2,26 +2,16 @@ using Decorator.Attributes;
 
 namespace Decorator.Demo.MessageTypes
 {
-	[Message("chat")]
-	public class Chat
-	{
-		[Position(0)]
-		public uint PlayerId { get; set; }
+    [Message("chat")]
+    public class Chat
+    {
+        [Position(0)]
+        public uint PlayerId { get; set; }
 
-		[Position(1)]
-		public string ChatMessage { get; set; }
+        [Position(1)]
+        public string ChatMessage { get; set; }
 
-		public override string ToString()
-			=> $"[{nameof(Chat)}]: {nameof(PlayerId)}: {PlayerId} {nameof(ChatMessage)}: {ChatMessage}";
-	}
-
-	[Message("chat")]
-	public class SendChat
-	{
-		[Position(0)]
-		public string ChatMessage { get; set; }
-
-		[Position(1)]
-		public uint ClientId { get; set; }
-	}
+        public override string ToString()
+            => $"[{nameof(Chat)}]: {nameof(PlayerId)}: {PlayerId} {nameof(ChatMessage)}: {ChatMessage}";
+    }
 }
