+++ Decorator.Demo/MessageTypes/ClientEvent.cs
@@ -2,23 +2,10 @@ using Decorator.Attributes;
 
 namespace Decorator.Demo
 {
-	public class ClientEvent : ClientExistsEvent
-	{
-		[Required]
-		[Position(2)]
-		public bool JoinState { get; set; }
-	}
-
-	[Repeatable]
-	[Message("ce")]
-	public class ClientExistsEvent
-	{
-		[Required]
-		[Position(0)]
-		public uint Id { get; set; }
-
-		[Optional]
-		[Position(1)]
-		public string Username { get; set; }
-	}
+    public class ClientEvent : ClientExistsEvent
+    {
+        [Required]
+        [Position(2)]
+        public bool JoinState { get; set; }
+    }
 }
