+++ Decorator.Tests/Util/ExampleHandlerClass.cs
@@ -1,0 +1,6 @@
+﻿namespace Decorator.Tests
+{
+    public class ExampleHandlerClass
+	{
+	}
+}
