+++ Decorator.Tests/Util/NeedsAttribute.cs
@@ -1,0 +1,10 @@
+﻿using Decorator.Attributes;
+
+namespace Decorator.Tests
+{
+    public class NeedsAttribute
+	{
+		[Position(0), Required]
+		public string WOWOW { get; set; }
+	}
+}
