+++ Decorator.Tests/Util/NoProperties.cs
@@ -1,0 +1,10 @@
+﻿using Decorator.Attributes;
+
+namespace Decorator.Tests
+{
+
+    [Message("noprop")]
+	public class NoProperties
+	{
+	}
+}
