+++ Decorator.Tests/Util/NonRepeatable.cs
@@ -1,0 +1,11 @@
+﻿using Decorator.Attributes;
+
+namespace Decorator.Tests
+{
+    [Message("rep")]
+	public class NonRepeatable
+	{
+		[Position(0)]
+		public int SomeValue { get; set; }
+	}
+}
