+++ Decorator.Tests/Util/NullType.cs
@@ -1,0 +1,9 @@
+﻿using Decorator.Attributes;
+
+namespace Decorator.Tests
+{
+    [Message(null)]
+	public class NullType
+	{
+	}
+}
